Добавил схемы сервисов. Изменил существующие схемы, для работы с сервисами.
